added render-logic.js
