+++ render-logic.js
@@ -1,0 +1,20 @@
+/**
+ *
+ * @param {number} playerIndex
+ * @param {number} tokenIndex
+ * @param {number} tokenPosition
+ * @return {string}
+ */
+export function getTokenContainerId(playerIndex, tokenIndex, tokenPosition) {
+    if (tokenPosition === -1) {
+        return `h-${playerIndex}-${tokenIndex}`
+    }
+
+    if (tokenPosition > 50) {
+        const safeIndex = tokenPosition % 50;
+        return `p${playerIndex}s${safeIndex}`
+    }
+
+    const markIndex = (tokenPosition + (13 * playerIndex)) % 52
+    return `m${markIndex}`
+}
