+++ test/index.html
@@ -7,6 +7,6 @@
 <div id="qunit"></div>
 <div id="qunit-fixture"></div>
 <script src="https://code.jquery.com/qunit/qunit-2.22.0.js"></script>
-<script src="test.js"></script>
+<script src="test.js" type="module"></script>
 </body>
 </html>
